pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Sergei Zuev <zuevsn86@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8.10"
django = "4.2.8"
python-dotenv = "^1.0.1"
dj-database-url = "^2.3.0"
gunicorn = "^23.0.0"
django-bootstrap5 = "^24.3"
whitenoise = "6.7.0"
django-filter = "^24.3"
rollbar = "^0.16.3"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"
pytest = "^8.3.3"
coverage = "^7.6.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
